% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formweight.R
\name{form.weight}
\alias{form.weight}
\title{form.weight}
\usage{
form.weight(formula, out = "formula", update = FALSE)
}
\arguments{
\item{formula}{A character string containing a chemical formula, e.g. "SiO2" or "Ca(OH)2"}

\item{out}{Defines the output format of the function. If unspecified or out = "formula" the function gives out the molecular weight for the given formula. If out = "element" the output is a data.frame which lists all elements and the respective weight multiplied by the number of atoms of this element within the compound.}

\item{update}{Binary (TRUE or FALSE). Sets whether the information on the elements and their atomic masses shall be updated from GitHub. The option might be outdated and the information not longer available online; if not set, the function uses an internal database.}
}
\value{
A numeric value or data.frame, as defined by the function parameter "out".
}
\description{
Calculates the molar weight of a chemical compound or the sum of the atomic weights for each element of a compound.
}
