% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmass.R
\name{xmass}
\alias{xmass}
\title{XRD mass}
\usage{
xmass(data)
}
\arguments{
\item{data}{A data.frame which contains the chemical compound formulas in the first column and the respective share in the second column.}
}
\value{
A data.frame which contains the elements found in the sample and their percentual share by mass.
}
\description{
Calculates the mass for the elements in a sample based on a data.frame which contains the compound formulas
and the respective percentual share of each compound.
}
